데마시아5/DEMACIA5/djangobackend/requirements.txt
alabaster==0.7.12
albumentations==0.4.6
appdirs==1.4.4
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.1.0
Babel==2.8.0
branca==0.3.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
dataframe==0.2.1.3
decorator==4.4.2
distlib==0.3.1
Django==3.1.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
docutils==0.16
drf-yasg==1.17.1
filelock==3.0.12
folium==0.10.1
idna==2.8
imageio==2.9.0
imagesize==1.2.0
imgaug==0.4.0
importlib-metadata==1.7.0
inflection==0.5.1
iniconfig==1.0.1
itypes==1.2.0
Jinja2==2.10.3
kiwisolver==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.5.0
MouseInfo==0.1.3
mysql-connector-python==8.0.21
mysqlclient==2.0.1
networkx==2.5
nose==1.3.7
numpy==1.17.3
openapi-codec==1.3.2
opencv-python==4.4.0.42
packaging==20.4
pandas==0.25.3
Pillow==7.2.0
plotly==4.9.0
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
Pygments==2.6.1
PyMsgBox==1.0.8
pyparsing==2.4.5
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.6
pytest==6.0.1
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-image==0.17.2
scipy==1.3.1
seaborn==0.9.0
Shapely==1.7.1
simplejson==3.17.2
six==1.13.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
tabulate==0.8.7
tifffile==2020.9.3
toml==0.10.1
tox==3.20.0
uritemplate==3.0.1
urllib3==1.25.7
virtualenv==20.0.31
zipp==3.1.0
